requirements.txt
Flask==2.0.2
flask-mongoengine==1.0.0
Flask-User==1.0.2.2
gunicorn==20.1.0
mongoengine==0.24.1
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.9